[GEOT-6846] MongoDB plugin refact FilterToMongo and FilterSplitter (#3403)
* [GEOT-6846] MongoDB plugin refact FilterToMongo and FilterSplitter

* review feedback applied
